clean order details code by splitting skeletons and components
