+++ frontend/src/modules/OrderModule/ReadOrderDetailsModule/components/OrderDetails/OrderDetailsSkeleton.tsx
@@ -1,0 +1,7 @@
+import { Skeleton } from "@mui/material";
+
+const OrderDetailsSkeleton = () => {
+    return <Skeleton />;
+};
+
+export default OrderDetailsSkeleton;
